feat: open login as modal overlay on landing page
Use Next.js intercepting routes + parallel slots to show the login
form as a modal when navigating from the landing page. Direct URL
access to /login still renders the full standalone page (now with a
'Back to home' link). Sign-out now redirects to / instead of /login.
