+++ changelog/en/2026-04-28-login-modal.mdx
@@ -1,0 +1,18 @@
+---
+version: "2026.04.30"
+date: "2026-04-28"
+title: "Sign in now opens as a popup over the home page"
+tags: ["improvement"]
+---
+
+Clicking **Sign in** or **Get started** on the home page now opens the login form as a popup right there — no more jumping to a separate page. You stay on the home page the whole time.
+
+Also, if you sign out, you now land on the home page instead of getting stuck on a blank login screen. And if you do end up on the login page directly (from a bookmark or link), there's a **Back to home** link in the top left.
+
+<div className="changelog-image-grid">
+
+![Before: the login page was a separate dead-end screen with no way back](/changelog/login-modal/before-login.png)
+
+![After: sign-in now appears as a popup over the landing page](/changelog/login-modal/after-modal.png)
+
+</div>
