+++ changelog/es/2026-04-28-login-modal.mdx
@@ -1,0 +1,18 @@
+---
+version: "2026.04.30"
+date: "2026-04-28"
+title: "Iniciar sesion ahora se abre como popup sobre la pagina principal"
+tags: ["improvement"]
+---
+
+Hacer clic en **Iniciar sesion** o **Empezar** en la pagina principal ahora abre el formulario de login como un popup justo ahi — ya no te lleva a una pagina separada. Te quedas en la pagina principal todo el tiempo.
+
+Ademas, si cierras sesion, ahora vuelves a la pagina principal en vez de quedarte atrapado en una pantalla de login en blanco. Y si llegas a la pagina de login directamente (desde un marcador o enlace), hay un enlace **Volver al inicio** arriba a la izquierda.
+
+<div className="changelog-image-grid">
+
+![Antes: la pagina de login era una pantalla separada sin forma de volver](/changelog/login-modal/before-login.png)
+
+![Despues: iniciar sesion aparece como popup sobre la pagina principal](/changelog/login-modal/after-modal.png)
+
+</div>
