+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lol-tracker",
-  "version": "2026.04.29",
+  "version": "2026.04.30",
   "private": true,
   "license": "AGPL-3.0-or-later",
   "scripts": {
