+++ src/app/@auth/default.tsx
@@ -1,0 +1,3 @@
+export default function AuthDefault() {
+  return null;
+}
