+++ src/components/login-modal.tsx
@@ -1,0 +1,36 @@
+"use client";
+
+import { Dialog as DialogPrimitive } from "@base-ui/react/dialog";
+import { useRouter } from "next/navigation";
+import * as React from "react";
+
+import { cn } from "@/lib/utils";
+
+/**
+ * A modal wrapper for the intercepted login route.
+ * Always renders open and navigates back on close (dismissing the overlay).
+ */
+export function LoginModal({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+
+  function handleOpenChange(open: boolean) {
+    if (!open) {
+      router.back();
+    }
+  }
+
+  return (
+    <DialogPrimitive.Root open onOpenChange={handleOpenChange}>
+      <DialogPrimitive.Portal>
+        <DialogPrimitive.Backdrop
+          className={cn(
+            "fixed inset-0 isolate z-50 bg-black/60 duration-100 supports-backdrop-filter:backdrop-blur-sm data-open:animate-in data-open:fade-in-0 data-closed:animate-out data-closed:fade-out-0",
+          )}
+        />
+        <DialogPrimitive.Popup className="fixed inset-0 z-50 flex items-center justify-center">
+          {children}
+        </DialogPrimitive.Popup>
+      </DialogPrimitive.Portal>
+    </DialogPrimitive.Root>
+  );
+}
